Overhaul error checking for load_object and save_object
More consistent useful codes, clearer implementation, ability to detect all
I/O errors.

Acknowledge recent changes, and promote save_object (back) to the VM spec
from the C SMite documentation.

+++ src/aux.h
@@ -90,7 +90,7 @@ int smite_byte_size(smite_WORD v); // return number of significant bytes in a sm
 #define INSTRUCTION_MAX_CHUNKS (((smite_word_bit + INSTRUCTION_CHUNK_BIT - 1) / INSTRUCTION_CHUNK_BIT))
 ptrdiff_t smite_encode_instruction_file(int fd, enum instruction_type type, smite_WORD v);
 ptrdiff_t smite_encode_instruction(smite_state *S, smite_UWORD addr, enum instruction_type type, smite_WORD v);
-int smite_decode_instruction_file(int fd, smite_WORD *val);
+ssize_t smite_decode_instruction_file(int fd, smite_WORD *val);
 int smite_decode_instruction(smite_state *S, smite_UWORD *addr, smite_WORD *val);
 
 // Object files
